refactor - separate component files
